refactor(engine): Make auth.clients top level
