+++ tests/unit/test_auth.py
@@ -3,7 +3,7 @@ import os
 import pytest
 from httpx import AsyncClient
 
-from tracecat.auth.clients import AuthenticatedAPIClient, AuthenticatedServiceClient
+from tracecat.clients import AuthenticatedAPIClient, AuthenticatedServiceClient
 from tracecat.config import TRACECAT__API_URL
 from tracecat.contexts import ctx_role
 from tracecat.types.auth import Role
